thinkAndFind and conceptQuest code review
