+++ littleTalkApp/templates/exercises/think_and_find.html
@@ -4,7 +4,7 @@
 <head>
     <meta charset="UTF-8">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Think &amp; Find Exercise</title>
+    <title>Think and Find Exercise</title>
     {% vite_hmr_client %}
     {% vite_asset 'frontend/src/style.css' %}
 </head>
