+++ static/js/exercises/css/concept-quest.css
@@ -1,1 +1,1 @@
-._settingsPanel_pxuib_1{display:flex;flex-direction:column;gap:1.5rem}._settingsGroup_pxuib_7{display:flex;flex-direction:column;gap:.85rem}._settingsHeading_pxuib_13{margin:0;font-size:1.125rem;font-weight:700}._settingsHint_pxuib_19{margin:0;color:#4b5563;font-size:.95rem;line-height:1.5}._selectionGrid_pxuib_26{display:grid;grid-template-columns:repeat(2,minmax(0,1fr));gap:.75rem}._selectionGridCompact_pxuib_32{display:grid;grid-template-columns:repeat(3,minmax(0,1fr));gap:.75rem}._optionsGrid_pxuib_38{display:flex;flex-wrap:wrap;justify-content:center;gap:1.25rem;width:100%;max-width:1200px;margin:0 auto}._optionsGrid_pxuib_38 button{width:clamp(180px,24vw,280px);height:clamp(180px,24vw,280px);padding:1.25rem}._optionsGrid_pxuib_38 img{width:clamp(130px,20vw,220px);height:clamp(130px,20vw,220px)}@media(max-width:640px){._selectionGrid_pxuib_26,._selectionGridCompact_pxuib_32{grid-template-columns:repeat(2,minmax(0,1fr))}._optionsGrid_pxuib_38{gap:.75rem}._optionsGrid_pxuib_38 button{width:clamp(130px,40vw,180px);height:clamp(130px,40vw,180px)}._optionsGrid_pxuib_38 img{width:clamp(90px,28vw,130px);height:clamp(90px,28vw,130px)}}
+._settingsPanel_xrdmn_1{display:flex;flex-direction:column;gap:1.5rem}._settingsGroup_xrdmn_7{display:flex;flex-direction:column;gap:.85rem}._settingsHeading_xrdmn_13{margin:0;font-size:1.125rem;font-weight:700}._settingsHint_xrdmn_19{margin:0;color:#4b5563;font-size:.95rem;line-height:1.5}._selectionGrid_xrdmn_26{display:grid;grid-template-columns:repeat(2,minmax(0,1fr));gap:.75rem}._selectionGridCompact_xrdmn_32{display:grid;grid-template-columns:repeat(3,minmax(0,1fr));gap:.75rem}._optionsGrid_xrdmn_38{display:flex;flex-wrap:wrap;justify-content:center;gap:1.25rem;width:100%;max-width:1200px;margin:0 auto}._optionsGrid_xrdmn_38 button{width:clamp(180px,24vw,280px);height:clamp(180px,24vw,280px);padding:1.25rem}._optionsGrid_xrdmn_38 img{width:clamp(130px,20vw,220px);height:clamp(130px,20vw,220px)}@media(max-width:640px){._selectionGrid_xrdmn_26,._selectionGridCompact_xrdmn_32{grid-template-columns:repeat(2,minmax(0,1fr))}._optionsGrid_xrdmn_38{gap:.75rem}._optionsGrid_xrdmn_38 button{width:clamp(130px,40vw,180px);height:clamp(130px,40vw,180px)}._optionsGrid_xrdmn_38 img{width:clamp(90px,28vw,130px);height:clamp(90px,28vw,130px)}}
